Consolidate confluence-scraper build process and cleanup
Migrated the build process to the main Makefile, enabling architecture-specific building logic. Removed duplicated .gitignore entries specific to confluence-scraper. Deleted the confluence-scraper directory which included outdated configuration, models, and internal files.

This change simplifies the project structure and centralizes the build configuration, reducing redundancy and preventing potential configuration drift between different build scripts.
