+++ .gitignore
@@ -25,3 +25,5 @@ __debug_*
 .env
 
 output/
+
+confluence-scraper
