--- cmd/confluence-scraper/.gitignore
@@ -1,2 +1,0 @@
-confluence-scraper
-output/
